Add rectangle toggle example
This commit adds a new example that shows a XOR pattern. A rectangle can
be superimposed on the background by pressing the space bar.

This example was created to demonstrate a bug in the Core Graphics
backend that causes the buffer to fade when calling set_buffer, instead
of applying it immediately.