Separate `Buffer` type, with a `present` method
This is similar to the API used in `winit`. It makes it impossible to
present without calling `buffer_mut`, and allows `Buffer` to contain
custom data and have custom drop behavior if `present` happens to not be
called.

This adds a utility type called `BorrowStack` to allow `Buffer` to
contain a reference to the surface and also a slice that may borrow from
it. There is a miri test to check this conforms to soundness rules.

The Orbital implementation here uses a no-copy method if the buffer has
the same size as the window, and a separate vec and copy otherwise.