fix(web): optimize and fix present_with_damage
This patch fixes the present_with_damage implementation for the web
backend. `CanvasRenderingContext2D`’s `putImageDate` was used
incorrectly.

As per Mozilla documentation:

> imageData
>
>   An ImageData object containing the array of pixel values.
>
> dx
>
>   Horizontal position (x coordinate) at which to place the image data
>   in the destination canvas.
>
> dy
>
>   Vertical position (y coordinate) at which to place the image data
>   in the destination canvas.
>
> dirtyX
>
>   Horizontal position (x coordinate) of the top-left corner from which
>   the image data will be extracted. Defaults to 0.
>
> dirtyY
>
>   Vertical position (y coordinate) of the top-left corner from which
>   the image data will be extracted. Defaults to 0.

As such, `dx` is different from `dirtyX` and `dy` is different from
`dirtyY`.

When the actual `ImageData` is as big as the canvas itself, `dx` and `dy`
should always be `0.0` and `0.0`, and only `dirtyX` and `dirtyY` are
actually representing the top-left coordinates for the dirty region.
The dirty region is relative to the image data.

This patch also optimizes the BGRX to RGBA conversion by only
creating a bitmap for the smallest region enclosing all the damaged
regions. This does not affect the behavior of `present`.