Use traits to define common backend interface
This still uses enums to dispatch to the backends, but has the `*Dispatch`
types implement the same traits. Nothing about the behavior or
performance should change.

If we require backends to implement the same methods, it seems good to
use traits for that. Any documentation on the interface implemented by
backends can be in one place. The traits can provide default
implementations where appropriate.

Hopefully this will help with more refactoring and features.