deps: Use `objc2`
Fixes #181.

This also prevents accessing the `NSView`/`NSWindow` from other threads than the main thread.

SoftBufferError is `#[non_exhaustive]`, so it's fine to add further
variants to it. I decided against doing that though, as the error is
platform-specific, and not one that the program can really handle in any
reasonable way (it's closer to a programmer error).