m: New multithreading strategy for Windows
This new strategy discards the old subclassing strategy, which led to deadlocks
when the main thread was blocked. Instead we instantiate a new thread
specifically for creating and destroying DC's. Then we set up a channel so that
all DCs are created and then destroyed on that thread.

I am open to suggestions as to ways to avoid allocating the new thread.

Signed-off-by: John Nunley <dev@notgull.net>